hat_detector_web/requirements.txt
Flask
opencv-python
azure-cognitiveservices-vision-computervision
msrest
gunicorn
